build: introduce LIBSOLVER environment variable
When building solvers.a while aiming at building CBMC with support for
IPASIR solvers, all projects that need SAT checkers would require to
link against the provided libipasir.a library. To make this library
visible for all subprojects at the same time, but furthermore allow the
user to choose which library to pick, the variable LIBSOLVER was
introduced, which is set to an empty value by default.

To get started with the default ipasir package, run the following commands
in the root directory of the repository:

make ipasir-build -C src
make -C src IPASIR=../../ipasir LIBSOLVER=$(pwd)/ipasir/libipasir.a