Don't stop when recursion found
Since we are finding a fixed point we do the same things we do for loops
when we find a recurssive function - we descend into it and carry on.

Previously we were never doing the transforms from the recurssive
function call to the start of the function meaning we would not realise
it can be called with multiple parameters. Further, we weren't doing the
transform from after the recursive function call onwards, meaning for
some inputs it would incorrectly conclude the function never returned.

Now we just do the transforms ignoring the fact it could be recursive.
For the constants domain this is OK as a fixed point will be found
quickly. For other domains we will need to implement some sort of
widening.

Included a test for recursion.