Don't make copies unnecessarily when merging
When we merge, we don't want to create a clone of the object if nothing
changes as this reduces the ability of the sharing map to optimize.

This simplifies the code for detecting whether a merge has changed
anything, as we can compare the pointers, if they are pointing at the
same thing, nothing has changed.

It complicates the code for the merges as we need to be careful not to
leak any abstract objects. We can't use the shared pointer wrapper as we
want to be able to use the this pointer (unless we can use
std::enable_shared_from_this?)