Implemented "last written location" tracking abstract objects
Each abstract_objectt now has an std::set<goto_programt::const_targett> ("last write locations") data structure attached to it.

For each assign instruction, whenever an object is assigned to a value, the last write location set for that object is cleared, and updated to contain the program location (specifically the const_targett.)

Furthermore, if the object is not a simple symbol (e.g. struct, array), then the high-level symbol is also updated to represent modification. (unless it's a pointer write, in which case only modified objects are updated.)

For writes where multiple objects are copied 'simultaneously' (within a single GOTO instruction, for example int a[2] = { 3, 4 } then all sub-elements of the written object are also updated.  So in this case, given current location is X, then last_written_locations will appoximately map a->X, a[0]->X, a[1]->X.  This applies recursively.

Merges perform an additional union between objects' last write location sets.

The last_written_locations information is automatically outputted if output() is called, it is represented by a @ [X] or a @[X, Y], etc.