Fixing the method to work with java_bytecode
The java bytecode entry point relied on things working a little
differently.

Firstly, the function name is transformed to allow simpler function
names to be passed in. As such the rebuild_entry_point cannot be
responsible for looking up the symbol. This in turn involved a new way
of identifying the mode to use (using the existing entry points mode)

Secondly, as the lanuages are destroyed after the parsing phase, we must
re-provide the message handler as this is used by the entry point
generation.

Finally, the java_bytecode_entry_point adds a new symbol to the symbol
table, so we remove all symbols that are in the scope of the original
entry point.