Add identifier and rename statement to java_new_array_data
This will be used as an alternative to java_new_array.
This is to distinguish the statement before and after symex:
java_new_array is used before symex, it is then transformed into some
instrtuction containing a java_new_array_data by symex.
Symex use to add a new java_new_array statement which was confusing.