Deal with generic methods
The signature of a method can start with an angle bracket if it is a
generic method on a non-generic type or a static generic method.
Previously this would cause us to return a nil type rather than a code
type. This information is currently ignored and instead we just skip
over the Formal Type parameters and proceed using normal type
evaluation.