Reverting method descriptor loading
The signature and the descriptor of a method do not match up (see java
spec: The signature and descriptor (§4.3.3) of a given method or
constructor may not correspond exactly, due to compiler-generated
artifacts. In particular, the number of TypeSignatures that encode
formal arguments in MethodTypeSignature may be less than the number of
ParameterDescriptors in MethodDescriptor.) Reverting this to descriptor
solves the current problem with passing tests.