Make get_array return array of unknown expr
This is in the case where super_get does not anything about the array we
give it.
This is necessary so that the string solver returns an array of the
correct size, if the size is the only thing we know about the string.

Add warning message in the case where the symbol is unknown to super_get