Call graph: optionally record callsites
This makes it easier for call-graph users to discover the particular instruction responsible
for a call-graph edge.

Ideally these would be recorded in-line with the map (i.e. change the graph type to
std::multimap<irep_idt, std::pair<locationt, irep_idt>>), but here we introduce an optional
auxiliary map to avoid breaking existing users.