Fixes wrong invocation order for static initializers
The Java Virtual Machine Specification, section 5.5, p. 359 defines that the
static initializers of both the parent class or interfaces of a given class must
be invoked **before** the class initializer of the class. They were
previously being invoked in the opposite order.

Two Regression tests added to enforce this behaviour.