Remove useless cmdlinet::clear() call
In cmdlinet this clears vectors that are about to die anyway, and because
it is called in a destructor it cannot dispatch to a child class. By
removing this call we can disabuse any extending child of the false
expectation that their derived clear() method will in fact be called.