Use UTF-16 conversion function in expr2java
newline and carriage return need to be printed as \n and \r in Java
code, respectively (as opposed to \u000a and \u000d). ' and \ need to be
escaped. The previous implementation was missing those cases. In
addition to that, a call to an existing function is a cleaner way of
writing this.