Fixes to exception handling behaviour
Corrected the ordering of exception catching and added tests to verify the
correct behaviour.  Previously, the universal catch would be the one on the
outermost try (should be the innermost try).
Also, a throw can never throw directly past this catch, so there is no need to
include these alternatives (they will appear in the rethrow of the finally).