Add type parameters in signatures as dependencies
This patch analyzes the generic signatures in a class file in order to load
generic type args as class dependencies. All those types are added to the
class_refs of the parser to complete the dependency information. Before, a Java
object type that was only used as a generic type argument was not necessarily
loaded.