Add call to cproverNondetInitialize in nondet-init
This is for the user to be able to provide invariants about the objects
in the java programs.
This can avoid false positives in the detection of violated properties.
For instance many classes in the JDK have a field `size` that will
always be positive, providing a cproverNondetInitialize method lets
CBMC know about that.