Convert --intervals and --non-null from being specific to general analysis.
This may change the text output slightly but the core functionality remains the same
for intervals.  Non-null only existed as an option and wasn't actually implemented,
so it is disabled for now, with the options resulting in a more helpful error message.

Also this deprecates static_analyzer.{h,cpp} as the only use in the main code-base
has been replaced by static_show_domain.