Class hierarchy: add DOT output, unit tests
This also adds a goto-instrument command to output the class hierarchy in text or
DOT format. This is currently not usable with vanilla cbmc because none of its tools
can produce GOTO binaries from Java class files, but out-of-tree tools can, and JBMC
will surely gain support for GBF or its successors in the near future.