Replace unsound struct-cast simplification
Previously the simplifier would simplify ((struct A)x) into (x) if x.type() and A
were structurally compatible (i.e. had the same named and typed fields). However this
isn't really a sound simplification, as it changes the type of the expression. I therefore
replace the rule with one that only applies within the common-case member expression, i.e.
simplifiying ((struct A)x).y -> x.y, which does not change the type of the outer expression.