Fix Java ternary-compare against NaN
This was supposed to check if either operand is NaN and respond appropriately,
but using float-eq to do the test was incorrect, as NaN == x is false for any x,
whether NaN or not. This patch uses the isnan exprt to do the test instead, and adds
a regression test.