Refactoring of substitute array access
We make substitute_array return an expression.
We add an argument to left-propagate values in the substitution.
With this, we can avoid having to first replace the array_with with an
array and then propagating to overwrite unknown values in the array.

This function is now implemented using expr_try_dynamic_cast, which is
better than checking ids:
it can perform additional checks than just the id, it also does the
check and convert in one function call.

We split the function in several parts each responsible to handle a
specific type, which makes the global structure clearer.