Reimplement remove-static-init-loops to avoid need to inspect GOTO program
This was supplying synthetic --unwind parameters for each loop found in the GOTO program,
but that required early access to the GOTO code, which will not be available under symex-
driven lazy loading. Therefore we replace that pass with a handler mechanism that allows
the driver program to provide its own callback to make unwinding decisions, and use it in
JBMC to ensure that enumeration type static initialisers are unwound sufficiently.