Insert stub static globals onto any incomplete ancestor class
Previously they were always created on either the class that was referred to,
or on some parent or grandparent; however, this was inappropriate if all those
classes were concrete (non-stub) types but some interface *was* a stub. In that
case an interface will be arbitrarily picked and given a stub field.