Java method converter: look for inherited static fields
Previously it assumed that static field references were resolved in the bytecode
(e.g. that a reference to A.somefield meant that A defines somefield). However in
fact they can be inherited, and especially now that stub fields can be created on
parent classes they are likely to appear there. As such the "get/putstatic" instructions
now search the class hierarchy for the field they actually refer to.