Correct string-max-input-length tests
Remove test from max_input_length

This remove test which was using the way max-input-length was limiting
the size of strings, even constant one.
The solver is now more permisives and the long strings would just get
troncated instead of leading to a solver contradiction.

For MemberTest, we add a new test.
We now have two tests with different values of the option, which should
give different results.