Get rid of default axioms for strings
These default axioms where too strict and where applied even for strings
that may never be created in the actual program.
This could for instance lead to problematic contradiction for a Java
object which could be casted either to a string or an int:
the axiom added there would apply both on the string and on the integer.